+++ packages/react-i18n/src/i18n.ts
@@ -1,5 +1,4 @@
 import {
-  getDateDiff,
   isLessThanOneHourAgo,
   isLessThanOneMinuteAgo,
   isLessThanOneWeekAgo,
@@ -7,9 +6,7 @@ import {
   isToday,
   isYesterday,
   TimeUnit,
-  Weekdays,
-} from '@shopify/javascript-utilities/dates';
-import {applyTimeZoneOffset} from '@shopify/dates';
+} from '@shopify/dates';
 import {memoize as memoizeFn} from '@shopify/function-enhancers';
 import {memoize} from '@shopify/decorators';
 import {languageFromLocale, regionFromLocale} from '@shopify/i18n';
@@ -424,16 +421,15 @@ export class I18n {
   }
 
   private humanizeDate(date: Date, options?: Intl.DateTimeFormatOptions) {
-    const {defaultTimezone} = this;
-    const timeZone =
-      options && options.timeZone ? options.timeZone : defaultTimezone;
-
     if (isLessThanOneMinuteAgo(date)) {
       return this.translate('humanize.now');
     }
 
     if (isLessThanOneHourAgo(date)) {
-      const minutes = getDateDiff(TimeUnit.Minute, date);
+      const now = new Date();
+      const minutes = Math.floor(
+        (now.getTime() - date.getTime()) / TimeUnit.Minute,
+      );
       return this.translate('humanize.minutes', {
         count: minutes,
       });
@@ -453,15 +449,12 @@ export class I18n {
     }
 
     if (isLessThanOneWeekAgo(date)) {
-      const localTimeZone = Intl.DateTimeFormat().resolvedOptions().timeZone;
-
-      const dateInTimeZone =
-        timeZone === undefined
-          ? date
-          : applyTimeZoneOffset(date, timeZone, localTimeZone);
-
+      const weekday = this.formatDate(date, {
+        ...options,
+        weekday: 'long',
+      });
       return this.translate('humanize.weekday', {
-        day: Weekdays[dateInTimeZone.getDay()],
+        day: weekday,
         time,
       });
     }